+++ spec/overcommit/git_config_spec.rb
@@ -22,4 +22,59 @@ describe Overcommit::GitConfig do
       end
     end
   end
+
+  describe '.hooks_path' do
+    subject { described_class.hooks_path }
+
+    context 'when not explicitly set' do
+      around do |example|
+        repo do
+          example.run
+        end
+      end
+
+      it 'returns the default hook path' do
+        expect(subject).to eq File.expand_path(File.join('.git', 'hooks'))
+      end
+    end
+
+    context 'when explicitly set to an empty string' do
+      around do |example|
+        repo do
+          `git config --local core.hooksPath ""`
+          example.run
+        end
+      end
+
+      it 'returns the default hook path' do
+        expect(subject).to eq File.expand_path(File.join('.git', 'hooks'))
+      end
+    end
+
+    context 'when explicitly set to an absolute path' do
+      around do |example|
+        repo do
+          `git config --local core.hooksPath /etc/hooks`
+          example.run
+        end
+      end
+
+      it 'returns the absolute path' do
+        expect(subject).to eq '/etc/hooks'
+      end
+    end
+
+    context 'when explicitly set to a relative path' do
+      around do |example|
+        repo do
+          `git config --local core.hooksPath my-hooks`
+          example.run
+        end
+      end
+
+      it 'returns the absolute path to the directory relative to the repo root' do
+        expect(subject).to eq File.expand_path('my-hooks')
+      end
+    end
+  end
 end